computer-use/requirements.txt
openai>=1.68.2
pyautogui>=0.9.54
Pillow>=11.1.0
python-dotenv>=1.0.0